+++ packages/aws-amplify-react/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.3-unstable.0"></a>
+## [1.0.3-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react@1.0.2...aws-amplify-react@1.0.3-unstable.0) (2018-07-20)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="1.0.2"></a>
+## [1.0.2](https://github.com/aws/aws-amplify/compare/aws-amplify-react@1.0.2-unstable.1...aws-amplify-react@1.0.2) (2018-07-19)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="1.0.2-unstable.1"></a>
+## [1.0.2-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react@1.0.1...aws-amplify-react@1.0.2-unstable.1) (2018-07-19)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="1.0.2-unstable.0"></a>
+## [1.0.2-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react@1.0.1...aws-amplify-react@1.0.2-unstable.0) (2018-07-19)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="1.0.1"></a>
+## [1.0.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react@1.0.1-unstable.4...aws-amplify-react@1.0.1) (2018-07-18)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="1.0.1-unstable.4"></a>
+## [1.0.1-unstable.4](https://github.com/aws/aws-amplify/compare/aws-amplify-react@1.0.1-unstable.3...aws-amplify-react@1.0.1-unstable.4) (2018-07-18)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="1.0.1-unstable.3"></a>
+## [1.0.1-unstable.3](https://github.com/aws/aws-amplify/compare/aws-amplify-react@1.0.1-unstable.2...aws-amplify-react@1.0.1-unstable.3) (2018-07-18)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="1.0.1-unstable.2"></a>
+## [1.0.1-unstable.2](https://github.com/aws/aws-amplify/compare/aws-amplify-react@1.0.1-unstable.1...aws-amplify-react@1.0.1-unstable.2) (2018-07-18)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="1.0.1-unstable.1"></a>
+## [1.0.1-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react@1.0.1...aws-amplify-react@1.0.1-unstable.1) (2018-07-18)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="1.0.1-unstable.0"></a>
+## [1.0.1-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react@1.0.1...aws-amplify-react@1.0.1-unstable.0) (2018-07-18)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
 <a name="0.1.55-unstable.1"></a>
 ## [0.1.55-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react@0.1.55-unstable.0...aws-amplify-react@0.1.55-unstable.1) (2018-07-03)